requirements.txt
mlflow
Flask==3.0.3
torch==2.4.0
torchvision==0.19.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
Pillow==10.3.0
scikit-learn==1.5.1
python-dotenv==1.0.1
